Initial support for parsing .jigdo file
Add machinery to parse data out of a .jigdo file. gzip-compressed files
are not yet supported. There is just enough support for parsing the
.jigdo file to allow for discovery of the associated .template file,
so long as the .jigdo file specifies an absolute path, or the .template
file is in the CWD.

pigdo now takes as its argument the path to the .jigdo file, rather
than the path to the .template file.

The parsing is somewhat permissive in that it only searches for known
section names and key = value pairs, rather than rejecting any unknown
non-comment text. The jigdo-file(1) manual page does not specify if
fields are mandatory or not; for now, selectively enforce that fields
be set based on what the bare minimum would be to reconstruct an image.