Add support for decoding Base64-encoded MD5 sums
Jigdo Base64 differs in a few ways from RFC 3548 (or RFC 2045, since
jigdo predates RFC 3548 by a bit) Base64 in a few ways:

    * Jigdo Base64 uses '-' rather than '+' to represent the digit 62
    * Jigdo Base64 uses '_' rather than '/' to represent the digit 63
    * Jigdo Base64 does not use '=', or any other character, to pad
      the end of encoded messages whose length is not a multiple of
      three bytes, since jigdo only uses its Base64 encoding for MD5
      checksums, which have a fixed length. This length happens to be
      a non-multiple of three, so attempting to decode a jigdo Base64
      encoded MD5 sum with a standard Base64 decoded (even one that
      takes the alternate alphabet into account) will result in a
      corrupted checksum.

Add the necessary machinery for decoding the Base64-encoded checksums
in .jigdo files, and plumb it through to the parsing of the .template
checksum in the [Image] section. No routine for *encoding* Base64 is
provided, since that's only needed during .jigdo file creation.