Sort jigdoFileInfo records by MD5 sum
The .template file identifies files only by their MD5 checksum, so that
is the only way they can be looked up from the records in the .jigdo
file's [Parts] section. Sort the parsed records by their MD5 checksum
to make it easier to look up the file associated with an MD5.