Enable searching for file by MD5
Add a function to search a jigdoFileData's files[] array to find files
matching an MD5 checksum. This is structured to potentially return
multiple results, since the jigdo-file(1) manual page says that .jigdo
files may contain multiple path mappings to the same MD5 checksum.

The function is static for now, but might be exported depending on how
the API between jigdo.c and the main program ends up evolving.