Enable -Wall and -Werror in CFLAGS
These probably should have been enabled from the beginning.

Add a cast to unsigned when using an element of a char * as an array
subscript to the Base64 decoder table, to avoid -Wchar-subscripts.
Changing the char * to an array of unsigned characters was problematic
due to using the string elsewhere with standard library functions that
accept plain (and potentially signed) char types, e.g. strlen(3).

Preprocess out a recently added static function that isn't being used
yet, to avoid an expected -Wunused-function warning.