Reconstruct data from .template file into the target file
The .template file contains a stream of compressed data storing the
data that did not come from a matched file. The stream may be broken
into multiple chunks, all of which put together are a concatenation of
the TEMPLATE_ENTRY_TYPE_DATA entries in the DESC table.

This change adds rudimentary decompression support for zlib compressed
data in the .template file. The same support will eventually be used
for bzip2 compressed data in .template file, and for decompressing
gzip-compressed .jigdo files as well.