Wire up fetching code
Plumb together all of the pieces so far to start actually reassembling
the target file from all parts, including ones that need to be fetched.
There are still a lot of things that need to be supported (additional
valid possibilities in .jigdo file format specification, abillity to
suggest additional mirrors and local copies, optional local cache,
checksum verification of fetched parts and of the completed file,
recording a log of the commitStatus for each part to allow resuming,
etc.), and a few outright ugly holes (e.g. need better handling of
persistent fetch errors), but pigdo is now complete enough to serve as
a minimal proof of concept, and can successfully reassemble files from
a .jigdo/.template file pair that fall within its current limitations.