Add /usr/local to the default include and linker search paths
FreeBSD installs software that is not part of the base system to
/usr/local by default, but does not add /usr/local/include to the
default header include search path, or /usr/local/lib to the default
linker search path. Explicitly set the relevant compiler and linker
flags in the makefile to allow GNU make to build out of the box on
FreeBSD and other similarly configured systems.