Verify file checksum after reconstruction
Add a function to calculate the MD5 checksum of a file using an open
file descriptor. Work on the checksum 1024 pages at a time, to avoid
mapping all of a potentially giant file into memory at once, and also
to avoid accidentally requesting a mapping with an offset that doesn't
fall on a page boundary.