Add support for gzip-compressed .jigdo files
zlib has a very convenient API that mimics stdio and transparently
handles both uncompressed and gzip-compressed data; unfortunately, the
.jigdo parsing code relies heavily upon getline(3), which does not have
a workalike in zlib's stdio-like API for gzip streams.

Instead of using e.g. gzread(), push the fopen() call into a wrapper
function that decompresses gzip into a temporary file, returning a
FILE * for the temporary file if the target file was gzipped, or a
FILE * to the original file if it was not gzipped.