Simplify decompressMemToMem() API
The original intent was to use the same function to handle zlib, bzip2,
and gzip decompression; however, the gzip decompression workflow (which
is what motivated the buffer resizing and unknown input size support in
the original decompressMemToMem() API) was sufficiently different from
the zlib and bzip2 use cases that MemToMem didn't turn out to be the
way to go for handling gzipped .jigdo files. Remove support for unknown
input sizes and output buffer resizing from the decompression routines
in decompress.c.