Add wrapper to transparently fopen() files fetched over network
It would be useful to be able to download .jigdo and .template files
from a remote location, rather than being limited to local files on
the filesystem. With that in mind, add a wrapper around fopen() and
a newly added fetchToFile() helper to transparently fetch non-local
resources to a temporary file, and return a handle to the fetched file
where applicable, or a direct handle when opening a local file.

This is suitable for use with template files as-is, but will require
adjustment to gunzopen() to be suitable for compressed .jigdo files.