Download files in reverse size order
Large files naturally take longer to download, so a few large files at
the end of the file list could end up starving the worker queue if all
of the smaller files have already finished, leaving several worker
threads idle as the large files complete.

In order to take better advantage of parallelism, sort the files in
reverse order by size, so that the largest files are all completed
while there is still a glut of files to keep the workers busy when
they complete, and download progressively smaller and smaller files
until completion.