Initialize libcurl earlier
Now that it's possible to load .jigdo and .template files with libcurl,
the call to fetch_init() (which wraps thinly around curl_global_init())
should be moved so that it occurs before the first potential non-init
call to libcurl. libcurl initializes itself lazily anyway, so it's not
a huge deal, but the initialization is not thread-safe. It doesn't
actually matter, since the fetching that might occur before the current
place where fetch_init() all happens withing the main thread, but move
the initialization anyway in case something changes later.