Resume partially reassembled files
If the output file already exists, assume that it's partially complete
from a previous reassembly attempt. Verify that content corresponding
to known files is correct, and mark any verified files as complete.
Since this makes it possible to resume without requiring a log, delete
the TODO item about a log file from the README. A log file can still
optimize resuming a particularly large file, as it would allow skipping
the verification of the individual file parts.

The stream of data from the .template file that does not correspond to
any subcomponent files is restored unconditionally, since the .template
file does not contain enough data to verify whether the data pieces
have been restored correctly (e.g. a checksum)