Move .jigdo handling code into a static library
The API will need a lot of refinement to make it actually sensible to
use, but the general idea is to separate the basic .jigdo processing
and download code from the command line parsing and parallelization
glue. No DSO target for now, since the ABI will likely undergo a lot
of churn for the foreseeable future.